[OPS-133 / #12]refactor: UserLikeRecipe domain 폴더로 위치 이동
